improve transformation of group roles
